Added a lot of logics

+++ FavouriteGamesBot/Service/ListMenuService.cs
@@ -12,10 +12,12 @@ namespace FavouriteGamesBot.Service;
 
 public class ListMenuService
 {
+    private IGamesRepository _gamesRepository;
     private IGamesListsRepository _gamesListsRepository;
 
-    public ListMenuService(IGamesListsRepository gamesListsRepository)
+    public ListMenuService(IGamesRepository gamesRepository, IGamesListsRepository gamesListsRepository)
     {
+        _gamesRepository = gamesRepository;
         _gamesListsRepository = gamesListsRepository;
     }
 
@@ -50,7 +52,6 @@ public class ListMenuService
 
         if (gamesList == null)
         {
-
             return new BotMessage(DialogsStringsStorage.PressButton, GetGamesListsAsKeyboard(transmittedData));
         }
         else
@@ -58,7 +59,7 @@ public class ListMenuService
             transmittedData.State = States.ListMenu.ClickActionButtonWithList;
             transmittedData.DataStorage.AddOrUpdate("listId", gamesList.Id);
 
-            return new BotMessage(DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose);
+            return new BotMessage(DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose, true);
         }
 
         throw new Exception("Неизвестная ошибка в ProcessClickOnInlineButtonUserLists");
@@ -108,120 +109,19 @@ public class ListMenuService
         throw new Exception("Неизвестная ошибка в ProcessNewListName");
     }
 
-    public BotMessage ProcessGameAdding(string textData, TransmittedData transmittedData)
-    {
-        string state = transmittedData.State;
-
-        if (state == States.GameMenu.InputGameName)
-        {
-            if (textData.Length > 25)
-            {
-                return new BotMessage(DialogsStringsStorage.GamePriceErrorInput, null);
-            }
-            else
-            {
-                transmittedData.State = States.GameMenu.InputPrice;
-                transmittedData.DataStorage.AddOrUpdate("gameTitle", textData);
-
-                return new BotMessage(DialogsStringsStorage.GamePriceInput, null);
-            }
-        }
-        else if (state == States.GameMenu.InputPrice)
-        {
-            if (int.TryParse(textData, out int num) == false)
-            {
-                return new BotMessage(DialogsStringsStorage.GamePriceErrorInput, null);
-            }
-            else
-            {
-                if (num > 0)
-                {
-                    transmittedData.State = States.GameMenu.InputRating;
-                    transmittedData.DataStorage.AddOrUpdate("gamePrice", num);
-
-                    return new BotMessage(DialogsStringsStorage.GameRatingInput, null);
-                }
-                else
-                {
-                    return new BotMessage(DialogsStringsStorage.GamePriceErrorInput, null);
-                }
-            }
-
-        }
-        else if (state == States.GameMenu.InputRating)
-        {
-
-            if (int.TryParse(textData, out int num) == false)
-            {
-                return new BotMessage(DialogsStringsStorage.GameRatingErrorInput, null);
-            }
-            else
-            {
-                if (num > -1 && num < 11)
-                {
-                    transmittedData.State = States.GameMenu.InputComment;
-                    transmittedData.DataStorage.AddOrUpdate("gameRating", num);
-
-                    return new BotMessage(DialogsStringsStorage.GameDescriptionInput, null);
-                }
-                else
-                {
-                    return new BotMessage(DialogsStringsStorage.GameRatingErrorInput, null);
-                }
-            }
-        }
-        else if (state == States.GameMenu.InputComment)
-        {
-            transmittedData.State = States.GameMenu.InputCheck;
-            transmittedData.DataStorage.AddOrUpdate("gameComment", textData);
-
-            return new BotMessage(DialogsStringsStorage.Confirmation, null);
-        }
-        else if (state == States.GameMenu.InputCheck)
-        {
-            if (textData == "да")
-            {
-                transmittedData.State = States.ListMenu.ClickActionButtonWithList;
-
-                string title = (string)transmittedData.DataStorage.Get("gameTitle");
-                int price = (int)transmittedData.DataStorage.Get("gamePrice");
-                int rating = (int)transmittedData.DataStorage.Get("gameRating");
-                string comment = (string)transmittedData.DataStorage.Get("gameComment");
-
-                GamesList gamesList = _gamesListsRepository.GetGamesListById((int)transmittedData.DataStorage.Get("listId"));
-
-                gamesList.Games.Add(new Game() { Title = title, Price = price, Rating = rating, Comment = comment });
-                FgbDbContext dbContext = new FgbDbContext();
-                dbContext.GamesLists.Update(gamesList);
-                dbContext.SaveChanges();
-
-                return new BotMessage(DialogsStringsStorage.GameAdded(gamesList.Games.Last()), InlineKeyboardMarkupStorage.ListMenuChoose);
-            }
-            else if (textData == "нет")
-            {
-                transmittedData.State = States.ListMenu.ClickActionButtonWithList;
-                return new BotMessage(DialogsStringsStorage.Confirmation, null);
-            }
-
-            return new BotMessage(DialogsStringsStorage.Confirmation, null);
-        }
-
-        return new BotMessage(DialogsStringsStorage.GameNameInput, null);
-    }
-
     public BotMessage ProcessClickOnDeleteListButton(string textData, TransmittedData transmittedData)
     {
         GamesList gamesList = _gamesListsRepository.GetGamesListById((int)transmittedData.DataStorage.Get("listId"));
 
-        if (textData == "да")
+        if (textData == BotButtonsStorage.GameMenu.Confirm.CallBackData)
         {
-            transmittedData.State = States.ListMenu.ClickActionButtonWithList;
-
             _gamesListsRepository.DeleteGamesList((int)transmittedData.DataStorage.Get("listId"));
 
-            return new BotMessage(DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose);
+            transmittedData.State = States.MainMenu.ClickOnInlineButton;
+
+            return new BotMessage(DialogsStringsStorage.MainMenu, InlineKeyboardMarkupStorage.MainMenuChoose);
         }
-        else if (textData == "нет")
+        else if (textData == BotButtonsStorage.GameMenu.Cancel.CallBackData)
         {
             transmittedData.State = States.ListMenu.ClickActionButtonWithList;
 
@@ -234,42 +134,49 @@ public class ListMenuService
     public BotMessage ProcessClickActionButtonWithList(string textData, TransmittedData transmittedData)
     {
         GamesList gamesList = _gamesListsRepository.GetGamesListById((int)transmittedData.DataStorage.Get("listId"));
-//todo
-        switch (textData)
+
+        if (textData == BotButtonsStorage.ListMenu.AddGame.CallBackData)
         {
-            case "AddGame":
-                transmittedData.State = States.GameMenu.InputGameName;
-                return new BotMessage(DialogsStringsStorage.GameNameInput, null);
-                break;
-            case "CheckGames":
-
-                break;
-            case "RenameList":
-                transmittedData.State = States.ListMenu.NewListName;
-
-                return new BotMessage(DialogsStringsStorage.NewListNameInput, null);
-                break;
-            case "ChangeListPrivacy":
-                transmittedData.State = States.ListMenu.ClickButtonChangePrivacy;
-
-                return new BotMessage(DialogsStringsStorage.ListPrivacyInfo(gamesList) + DialogsStringsStorage.ListPrivacy, InlineKeyboardMarkupStorage.ChooseListPrivacy);
-
-                break;
-            case "DeleteList":
-                transmittedData.State = States.ListMenu.ClickOnDeleteListButton;
-
-                return new BotMessage(DialogsStringsStorage.ListDeletedConfirmation, null);
-                break;
-            case "BackToLists":
-                transmittedData.State = States.ListMenu.ClickOnInlineButtonUserLists;
-                transmittedData.DataStorage.Delete("listId");
-
-                return new BotMessage(DialogsStringsStorage.MyLists, GetGamesListsAsKeyboard(transmittedData));
-                break;
-            default:
-                return new BotMessage(DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose);
-                break;
+            transmittedData.State = States.GameMenu.InputTitle;
+            return new BotMessage(DialogsStringsStorage.GameNameInput, null);
         }
+        else if (textData == BotButtonsStorage.ListMenu.CheckGames.CallBackData)
+        {
+            if (((List<Game>)gamesList.Games).Count == 0)
+                return new BotMessage(DialogsStringsStorage.GamesAreNull + "\n\n" + DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose);
+            else
+            {
+                transmittedData.State = States.GameMenu.ClickOnInlineButtonListGames;
+                return new BotMessage(DialogsStringsStorage.GamesInList, ReplyKeyboardMarkupStorage.CreateKeyboardGames((List<Game>)gamesList.Games));
+            }
+        }
+        else if (textData == BotButtonsStorage.ListMenu.RenameList.CallBackData)
+        {
+            transmittedData.State = States.ListMenu.NewListName;
+
+            return new BotMessage(DialogsStringsStorage.NewListNameInput, null);
+        }
+        else if (textData == BotButtonsStorage.ListMenu.ChangeListPrivacy.CallBackData)
+        {
+            transmittedData.State = States.ListMenu.ClickButtonChangePrivacy;
+
+            return new BotMessage(DialogsStringsStorage.ListPrivacyInfo(gamesList) + DialogsStringsStorage.ListPrivacy, InlineKeyboardMarkupStorage.ChooseListPrivacy);
+        }
+        else if (textData == BotButtonsStorage.ListMenu.DeleteList.CallBackData)
+        {
+            transmittedData.State = States.ListMenu.ListInputDeletingConfirmation;
+
+            return new BotMessage(DialogsStringsStorage.ListDeletedConfirmation, InlineKeyboardMarkupStorage.GameConfirmation);
+        }
+        else if (textData == BotButtonsStorage.ListMenu.BackToLists.CallBackData)
+        {
+            transmittedData.State = States.ListMenu.ClickOnInlineButtonUserLists;
+            transmittedData.DataStorage.Delete("listId");
+
+            return new BotMessage(DialogsStringsStorage.MyLists, GetGamesListsAsKeyboard(transmittedData));
+        }
+
+        return new BotMessage(DialogsStringsStorage.ChoosedList(gamesList), InlineKeyboardMarkupStorage.ListMenuChoose);
 
         throw new Exception("Неизвестная ошибка в ProcessClickActionButtonWithList");
     }
@@ -293,7 +200,7 @@ public class ListMenuService
             }
         }
 
-        return new ReplyKeyboardMarkup(rows.ToArray());
+        return new ReplyKeyboardMarkup(rows.ToArray()) { ResizeKeyboard = true };
     }
 
 }
